extracted monkey patching
