+++ src/js/abstract/base-component.js
@@ -9,6 +9,7 @@ import lifecycleLogger from './hocs/lifecycle-logger';
 
 import PropertyExistsException from './property-exists-exception';
 import withContext from './hocs/with-context';
+import withMonkeyPatches from './hocs/with-monkey-patches';
 
 const THROWED_ERROR = 'throwed';
 const PROPERTY_WHITELIST = ['title', 'checked', 'disabled'];
@@ -57,7 +58,7 @@ const getId = (nodeName) => {
  * </axa-example>
  * ```
  */
-export default withContext(class BaseComponent extends HTMLElement {
+export default withContext(withMonkeyPatches(class BaseComponent extends HTMLElement {
   constructor(styles = '', template) {
     super();
 
@@ -441,78 +442,6 @@ export default withContext(class BaseComponent extends HTMLElement {
   didRenderCallback(initial) {} // eslint-disable-line
 
   /**
-   * Monkey patch `innerText` API to re-rendering.
-   *
-   * @param {String} text
-   */
-  set innerText(text) {
-    if (!this._hasTemplate || !this._hasRendered) {
-      super.innerText = text;
-      return;
-    }
-
-    const textNode = document.createTextNode(text);
-
-    this._lightDOMRefs = [textNode];
-
-    this.render();
-  }
-
-  /**
-   * Monkey patch `textContent` API to re-rendering.
-   *
-   * @param {String} text
-   */
-  set textContent(text) {
-    if (!this._hasTemplate || !this._hasRendered) {
-      super.textContent = text;
-      return;
-    }
-
-    const textNode = document.createTextNode(text);
-
-    this._lightDOMRefs = [textNode];
-
-    this.render();
-  }
-
-  /**
-   * Monkey patch `innerHTML` API to re-rendering.
-   *
-   * @param {String} html
-   */
-  set innerHTML(html) {
-    if (!this._hasTemplate || !this._hasRendered) {
-      super.innerHTML = html;
-      return;
-    }
-
-    const div = document.createElement('div');
-
-    div.innerHTML = html;
-
-    this._lightDOMRefs = Array.from(div.children);
-
-    this.render();
-  }
-
-  /**
-   * Monkey patch `appendChild` API to re-rendering.
-   *
-   * @param {Element} node
-   */
-  appendChild(node) {
-    if (this._isMorphing || !this._hasTemplate || !this._hasRendered) {
-      super.appendChild(node);
-      return;
-    }
-
-    this._lightDOMRefs.push(node);
-
-    this.render();
-  }
-
-  /**
    * Only morph children of current custom element, not any other custom element.
    *
    * @returns {boolean}
@@ -527,4 +456,4 @@ export default withContext(class BaseComponent extends HTMLElement {
       return v.toString(16);
     });
   }
-});
+}));
